refactor: split watches into modules + WatchStore handle+channel
